Upgrade senScore

+++ src/BaseDeDonnee/BDD.java
@@ -81,7 +81,7 @@ public class BDD implements Parametre{
     }
 
     /*      Fonction qui permet d'inserer une ligne de score*/
-    protected static void insertLigneScore(String pseudoInsert, int mvt, int score, int chrono, Connection co){
+    protected static void insertLigneScore(String pseudoInsert, int mvt, int score, double chrono, Connection co){
                 /*  Variables */
         String pseudo = "'"+pseudoInsert+"'";
         boolean firstStrike = true; //c'est notre première partie
